+++ Sources/ListInserter/Inserter.swift
@@ -33,7 +33,6 @@ public class Inserter<S: Sectionable> {
         shouldInsertItems: @escaping () -> Bool = { true },
         itemInfoIdGenerator: @escaping () -> String = { UUID().uuidString }
     )
-
     {
         self.insertionRequests = insertionRequests
         self.shouldInsertItems = shouldInsertItems
@@ -392,113 +391,4 @@ public class Inserter<S: Sectionable> {
             lhs.identifier == rhs.identifier
         }
     }
-}
-
-public protocol Sectionable<SectionIdentifier, Value, Embed>: Identifiable, Hashable {
-    associatedtype SectionIdentifier where SectionIdentifier: Hashable
-    associatedtype Value where Value: ValueKindIdentifiable, Value: Hashable
-    associatedtype Embed
-    
-    var id: SectionIdentifier { get }
-    var items: [Item<Value, Embed>] { get set }
-}
-
-public protocol SectionableInitable: Sectionable {
-    init(items: [Item<Value, Embed>])
-}
-
-public struct Section<SectionIdentifier, Value, Embed> : Sectionable
-where
-SectionIdentifier: Hashable,
-Value: ValueKindIdentifiable,
-Value: Hashable
-{
-    public var id: SectionIdentifier
-    
-    public var items: [Item<Value, Embed>]
-    
-    public init(sectionIdentifer: SectionIdentifier, items: [Item<Value, Embed>]) {
-        self.id = sectionIdentifer
-        self.items = items
-    }
-}
-
-public struct NoSection<Value, Embed> : SectionableInitable
-where
-Value: ValueKindIdentifiable,
-Value: Hashable
-{
-    public var id: String
-    
-    public var items: [Item<Value, Embed>]
-    
-    public init(items: [Item<Value, Embed>]) {
-        self.id = UUID().uuidString
-        self.items = items
-    }
-}
-
-extension Item: Identifiable where Value: Identifiable, Value.ID == String {
-    public var id: String {
-        switch self {
-        case let .inserted(info):
-            return info.id
-        case let .value(value):
-            return value.id
-        }
-    }
-}
-
-public enum Item<Value, Embed>: Hashable, ItemKindIdentifiable
-    where
-    Value: Hashable,
-    Value: ValueKindIdentifiable
-{
-    public var itemKindId: ItemKind<Value> {
-        switch self {
-        case .inserted:
-            return .inserted
-        case let .value(value):
-            return .value(value)
-        }
-    }
-
-    case inserted(InsertedItemInfo<Embed>)
-    case value(Value)
-}
-
-public struct InsertedItemInfo<Embed>: Hashable, Identifiable {
-    public var id: String
-
-    public let embed: Embed
-
-    public init(embed: Embed, id: String) {
-        self.embed = embed
-        self.id = id
-    }
-
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(id)
-    }
-
-    public static func == (lhs: InsertedItemInfo, rhs: InsertedItemInfo) -> Bool {
-        lhs.id == rhs.id
-    }
-}
-
-public protocol ItemKindIdentifiable {
-    associatedtype ValueKind where ValueKind: ValueKindIdentifiable
-    
-    var itemKindId: ItemKind<ValueKind> { get }
-}
-
-public enum ItemKind<ValueKind: ValueKindIdentifiable>: Hashable {
-    case inserted
-    case value(ValueKind)
-}
-
-public protocol ValueKindIdentifiable: Hashable {
-    associatedtype H where H: Hashable
-    
-    var valueKind: H { get }
 }
